pkg/rag-brains/pyproject.toml
[tool.poetry]
name = "rag-brains"
version = "0.1.0"
description = "Shared RAG pipeline (chat, retriever) for CoW AI tools"
authors = ["José Ribeiro <jose@bleu.studio>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.2"
langchain-openai = "^0.1.14"
langchain-text-splitters = "^0.2.2"
langchain-community = "^0.2.6"
faiss-cpu = "^1.8.0.post1"
langchain-anthropic = "^0.1.19"
python-dotenv = "^1.0.1"
cow-core = { path = "../cow-core", develop = true }
tqdm = "^4.66.5"
langchain-core = "~0.2"
aiocache = "^0.12.3"
google-generativeai = "^0.8.0"
json-repair = "^0.24"
sentence-transformers = "^5.2.3"
pyyaml = "^6.0.2"

[tool.poetry.group.dev.dependencies]
ruff = "^0.5.0"
pytest = "^8.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
